refactor: consolidate helpers
