--- packages/repack/src/commands/common/adaptFilenameToPlatform.ts
@@ -1,10 +1,0 @@
-import os from 'node:os';
-
-const isWindows = os.platform() === 'win32';
-
-export const adaptFilenameToPlatform = (filename: string) => {
-  if (isWindows) {
-    return filename.replace(/\\/g, '/');
-  }
-  return filename;
-};
