+++ packages/repack/src/commands/common/index.ts
@@ -1,10 +1,10 @@
-export * from './adaptFilenameToPlatform.js';
 export * from './getDevMiddleware.js';
 export * from './getMimeType.js';
-export * from './resolveProjectPath.js';
-export * from './runAdbReverse.js';
 export * from './parseUrl.js';
 export * from './resetPersistentCache.js';
+export * from './resolveProjectPath.js';
+export * from './runAdbReverse.js';
+export * from './setupEnvironment.js';
 export * from './setupInteractions.js';
 export * from './setupStatsWriter.js';
