+++ packages/repack/src/commands/webpack/bundle.ts
@@ -1,5 +1,5 @@
 import webpack, { type Configuration } from 'webpack';
-import { CLIError } from '../common/cliError.js';
+import { CLIError } from '../../helpers/index.js';
 import { makeCompilerConfig } from '../common/config/makeCompilerConfig.js';
 import {
   normalizeStatsOptions,
