+++ packages/repack/src/helpers/helpers.ts
@@ -1,3 +1,21 @@
+import os from 'node:os';
+import type { Compiler as RspackCompiler } from '@rspack/core';
+import type { Compiler as WebpackCompiler } from 'webpack';
+
+// reference: https://github.com/web-infra-dev/rspack/discussions/2640
+export function isRspackCompiler(compiler: RspackCompiler | WebpackCompiler) {
+  return 'rspackVersion' in compiler.webpack;
+}
+
+export function isTruthyEnv(env: string | undefined): boolean {
+  return !!env && env !== 'false' && env !== '0';
+}
+
+export function adaptFilenameToPlatform(filename: string): string {
+  if (os.platform() !== 'win32') return filename;
+  return filename.replace(/\\/g, '/');
+}
+
 interface MoveElementBeforeConfig<T> {
   beforeElement: string | RegExp;
   elementToMove: string | RegExp;
