+++ packages/repack/src/plugins/DevelopmentPlugin.ts
@@ -7,8 +7,7 @@ import type {
 } from '@rspack/core';
 import ReactRefreshPlugin from '@rspack/plugin-react-refresh';
 import type { Compiler as WebpackCompiler } from 'webpack';
-import { isRspackCompiler } from './utils/isRspackCompiler.js';
-import { moveElementBefore } from './utils/moveElementBefore.js';
+import { isRspackCompiler, moveElementBefore } from '../helpers/index.js';
 
 const [reactRefreshEntryPath, reactRefreshPath, refreshUtilsPath] =
   ReactRefreshPlugin.deprecated_runtimePaths;
